Increase accuracy
Command: MPC_CALIBRATE HEATER=extruder TARGET=230
Done on an Ender 3 V2 Neo with 40 Watt heating element. Copper heat block.

Original:
0% fan average power: 21.81 W
50% fan average power: 20.74 W
100% fan average power: 21.01 W

Just 0.1 temperature target:
0% fan average power: 20.86 W
50% fan average power: 20.68 W
100% fan average power: 20.50 W

heat soak and stabilization:
0% fan average power: 20.65 W
50% fan average power: 20.61 W
100% fan average power: 21.53 W

0.1 target, heat soak and stabilization:
0% fan average power: 20.23 W
50% fan average power: 21.01 W
100% fan average power: 21.97 W

The 0.1 degree target method was better in some ways but it did not have the desired effect. That is because the extruder isn't getting heat soaked.
Heat soaking and stabilization is way better, though it still is not perfect. (Perhaps it is not heat soaked enough? We might need to play around with the 0.05.)
Both methods seem to work best but it comes at a significant time cost. Given that you should basically run it and forget it I would say that is worth it.

This change has flaws. Certain hardware might not be able to stabilize properly and that will cause swings too high for the 0.1 degree target. Therefore the 0.1 degree should probably be configurable and maybe not the default.

An alternative would be to build another function to keep track of the PWM of the heater and wait for that to stabilize. It probably wouldn't matter much in terms of time though.