Add wait_stable to MPC
The function waits for the temperature to cycle or for the temperature to be stable over a longer time.
This should ensure a more stable reading on the power usage.

As far as I can tell this is the best of both worlds. It will allow heaters that can't get a very stable output to continue and also heaters that do have a stable output to continue.